refactor(comps): modularize chart
